ocrs-erc-code/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
flasgger==0.9.7.1
mysql-connector-python==8.2.0
bcrypt==4.1.2
python-dotenv==1.0.0
email-validator==2.1.0
jsonschema==4.20.0
python-json-logger==2.0.7
pytest==8.0.0
pytest-flask==1.3.0
pytest-cov==4.1.0
requests==2.31.0
